% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R, R/estimate2.R
\name{control}
\alias{control}
\alias{control}
\title{Fitting Control.}
\usage{
control(qpoints = 30, init = NULL, predict = TRUE, adaptive = c("noad",
  "ad"))

control(qpoints = 30, init = NULL, predict = TRUE, adaptive = c("noad",
  "ad"))
}
\description{
Set the relvm fitting control parametes.

Set the relvm fitting control parametes.
}
