% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnll_cpp}
\alias{pnll_cpp}
\title{Prediction negLogLik Function}
\usage{
pnll_cpp(fv, score_row, wts_row, err, mu, fl)
}
\description{
Prediction negLogLik Function used in in pred function in pred.R
}
