% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Group Score Prediction}
\usage{
pred(score_tbl, wts_tbl, pms)
}
\arguments{
\item{score_tbl}{A matrix of standadized measure scores.}

\item{wts_tbl}{A matrix of weights.}

\item{pms}{A list or a data frame containing mu, fl, and err.}
}
\description{
Group Score Prediction
}
