% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_groups.R
\name{pred_single_measure_group}
\alias{pred_single_measure_group}
\title{Group score prediction from a single group}
\usage{
pred_single_measure_group(group, alldf, pars, mtbl)
}
\arguments{
\item{group}{The measure group specified.}

\item{alldf}{A data frame merged from both standardized measure scores and
measure weights.}

\item{pars}{The parameters in a data frame, including factor loading, err and
offset, etc.}

\item{mtbl}{A measure table calculated from rstarating::create_measure_tbl.}
}
\value{
relvm class object, which includes the predicted group scores and the
  corresponding standard deviations.
}
\description{
Predict/calculate a group score from a single group.
}
