% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{predict.relvm}
\alias{predict.relvm}
\title{Group Score Prediction}
\usage{
\method{predict}{relvm}(object, newdata, level = 0.95)
}
\arguments{
\item{pms}{A named vector of parameters to be optimized.}

\item{score_tbl}{The data table.}

\item{wts_tbl}{The corresponding weigting table.}
}
\value{
A data.frame containing pred and stderr. pred is the predicted group score.
  score.
}
\description{
Predict the group scores (random effect latent variabl) from the extimated
parameters
}

