% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_groups.R
\name{predict.relvms}
\alias{predict.relvms}
\title{Multiple Group score}
\usage{
\method{predict}{relvms}(x, newpars = NULL, newdata = NULL, level = 0.95)
}
\arguments{
\item{x}{An object, whose class attribute is "relvms", i.e. multiple group of
relvm. see also relvm function.}

\item{newpars}{The matrix of modified parameters (factor loadings, etc.).}

\item{newdata}{The new data table.}

\item{level}{The level of confident interval. This parameter may not be used
any more.}
}
\value{
A updated object.
}
\description{
Predict the group scores (random effect latent variabl) from the extimated
parameters
}
\examples{
fit <- relvm(mstbl(cmsdata::cms_star_rating_input_2017dec))
pred<- predict(fit,newpars=fit$groups$pars) # update the newpars.
star<-rating(x=pred$groups$summary_score,method="rclus2",score_col="sum_score",iter.max=5000)


}
