% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_adaptive2.R
\name{relvm_adaptive_single2}
\alias{relvm_adaptive_single2}
\title{Estimation Of The Random Effect Latent Variable Model Parameters}
\usage{
relvm_adaptive_single2(group, df, qpoints, inits, predict, adaptive)
}
\arguments{
\item{qpoints}{The numbe of the quadrature points.}

\item{inits}{Initial values for mu, fl, and err term in a list. fl is the
factor loading. They will be initialized generally if it is null. The
default is a list with for all mu and one for others.}

\item{predict}{The default is TRUE.}

\item{mstbl_std}{The standized measure score table.}

\item{wts_tbl}{The measure score weight table.}
}
\value{
An object of S3 class "relvm" with estimated parametes.
}
\description{
Estimate the random effect latent variable model
}
