% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{relvm_single}
\alias{relvm_single}
\title{Estimation Of The Random Effect Latent Variable Model Parameters}
\usage{
relvm_single(group, df = alldf, qpoints = 30, init = NULL,
  predict = TRUE)
}
\arguments{
\item{qpoints}{The numbe of the quadrature points.}

\item{init}{Initial values for mu, fl, and err term in a list. fl is the
factor loading. They will be initialized generally if it is null. The
default is a list with for all mu and one for others.}

\item{predict}{The default is TRUE.}

\item{mstbl_std}{The standized measure score table.}

\item{wts_tbl}{The measure score weight table.}
}
\value{
An object of S3 class "relvm" with estimated parametes.
}
\description{
Estimate the random effect latent variable model
}

