% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R, R/estimate3.R
\name{relvm_single3}
\alias{relvm_single3}
\alias{relvm_single3}
\title{Estimation Of The Random Effect Latent Variable Model Parameters}
\usage{
relvm_single3(group, df, qpoints, init, predict, adaptive)

relvm_single3(group, df, qpoints, init, predict, adaptive)
}
\arguments{
\item{qpoints}{The numbe of the quadrature points.}

\item{init}{Initial values for mu, fl, and err term in a list. fl is the
factor loading. They will be initialized generally if it is null. The
default is a list with for all mu and one for others.}

\item{predict}{The default is TRUE.}

\item{mstbl_std}{The standized measure score table.}

\item{wts_tbl}{The measure score weight table.}

\item{mstbl_std}{The standized measure score table.}

\item{wts_tbl}{The measure score weight table.}

\item{init}{Initial values for mu, fl, and err term in a list. fl is the
factor loading. They will be initialized generally if it is null. The
default is a list with for all mu and one for others.}

\item{predict}{The default is TRUE.}
}
\value{
An object of S3 class "relvm" with estimated parametes.

An object of S3 class "relvm" with estimated parametes.
}
\description{
Estimate the random effect latent variable model

Estimate the random effect latent variable model
}
