% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{relvm_single_true}
\alias{relvm_single_true}
\title{Estimation Of The Random Effect Latent Variable Model Parameters}
\usage{
relvm_single_true(group, df, inits, predict)
}
\arguments{
\item{group}{A measure group name.}

\item{df}{The standardized measure score and measure weight table (alldf).}

\item{inits}{Initial values for mu, fl, and err term in a list. fl is the
factor loading. They will be initialized generally if it is null. The
default is a list with for all mu and one for others.}

\item{predict}{The default is TRUE.}
}
\value{
An object of S3 class "relvm" with estimated parametes.
}
\description{
Estimate the random effect latent variable model
}
