% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R, R/relvm-package.r
\docType{package}
\name{relvm}
\alias{relvm}
\alias{relvm}
\alias{relvm-package}
\title{Multiple Estimation Of The Random Effect Latent Variable Model Parameters}
\usage{
relvm(object, groups = NULL, fit = list(init = NULL), file = NULL)
}
\arguments{
\item{object}{A mstbl object.}

\item{groups}{A vector of measure group names. The default is NULL, in which
case a vector of all groups will be generated accordingly.}

\item{fit}{A list of fitting parameters. \itemize{ \item init: Initial values for mu, fl, and err
term in a list. fl is the factor loading. They will be initialized
generally if it is null. The default is a list with for all mu and one for
others. \item predict: The default is TRUE.}}

\item{file}{A rds file name to cache the results on hard drive.}
}
\value{
An list of S3 object of class "relvms" with estimated parametes.
}
\description{
Estimate multiple measure groups of the random effect latent variable model

relvm.
}
\examples{
# require(rstarating)
# To fit just one group: 'outcome_mort'
fit <- relvm(mstbl(cms_star_rating_input_2017dec),groups="outcome_mort")

}
\seealso{
\code{\link{mstbl}}
}
